Start breaking up scripts
With how much can change in the time between computer setups, I'm
finding it easier to make a lot of little mini-scripts that I can run in
different orders as I find things that need fixing.

Also going to remove setup for Linux/WSL as I don't use that anymore.